requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
langchain-openai>=0.2.0

# Google Cloud
google-cloud-bigquery>=3.25.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0

# Data Validation
pydantic>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Testing
requests>=2.31.0
